+++ src/findTz.js
@@ -1,6 +1,6 @@
 'use strict'
 const Spacetime = require('./spacetime');
-const timezones = require('../data');
+// const timezones = require('../data');
 
 exports.whereIts = function(a, b) {
   let start = new Spacetime(null);
@@ -15,7 +15,7 @@ exports.whereIts = function(a, b) {
 
   let startHour = start.hour();
   let endHour = end.hour();
-  let tzs = Object.keys(timezones).filter(tz => {
+  let tzs = Object.keys(start.timezones).filter(tz => {
     let m = new Spacetime(null, tz);
     let hour = m.hour();
     //do 'calendar-compare' not real-time-compare