+++ test/immutable.test.js
@@ -5,12 +5,10 @@ const spacetime = require('../src/immutable');
 const day0 = spacetime.now();
 const today = day0.format('nice');
 
-test('clone still works', t=> {
+test('clone still works', t => {
   t.plan(2)
   var day1 = day0.clone()
-  console.log(day0.format('nice'))
   t.ok(day0.format('nice') === day1.format('nice'), 'eq')
-  console.log(day1.add(3, 'days').format('nice'))
   t.ok(day0.format('nice') === day1.format('nice'), 'eq')
 })
 
@@ -34,7 +32,7 @@ test('.hour does not mutate', t => {
 
 test('.date does not mutate', t => {
   t.plan(1);
-  let tmrw = day0.date([2018,1,1]).format('nice');
+  let tmrw = day0.date(1).month(1).year(2018).format('nice');
   t.ok(today === day0.format('nice'), 'not mutated');
 });